tests/unit/pages/payment-confirmation/payment-confirmation.test.ts
import { mocked } from 'ts-jest/utils';
import { PaymentConfirmationController } from '@pages/payment-confirmation/payment-confirmation';
import { PageNames } from '@constants';
import { Target, TCNRegion, TestType } from '../../../../src/domain/enums';
import {
  CRMBookingStatus,
  CRMNsaStatus, CRMOrigin, CRMProductNumber, CRMRemit,
} from '../../../../src/services/crm-gateway/enums';
import { BusinessTelemetryEvents, logger, getCreatedBookingIdentifiers } from '../../../../src/helpers/index';
import { PaymentStatus } from '../../../../src/services/payments/enums';
import { Candidate as PaymentCandidate, paymentGateway } from '../../../../src/services/payments/payment-gateway';
import { NsaBookingDetail } from '../../../../src/services/crm-gateway/interfaces';
import { store } from '../../../../src/services/session';
import { CrmServerError } from '../../../../src/domain/errors/crm';

jest.mock('../../../../src/services/payments/payment-gateway');
const mockPaymentGateway = mocked(paymentGateway, true);

jest.mock('../../../../src/services/session');
const mockStore = mocked(store, true);

const mockEmailContent = { subject: 'subject', body: 'body' };
jest.mock('../../../../src/services/notifications/content/builders', () => ({
  buildBookingConfirmationEmailContent: () => mockEmailContent,
}));

describe('Payment confirmation controller', () => {
  let controller: PaymentConfirmationController;
  let req: any;
  let res: any;

  const mockNotificationsGateway = {
    sendEmail: jest.fn(),
  };

  const mockBookingService = {
    completeBooking: jest.fn(),
    unreserveBooking: jest.fn(),
  };

  const mockCrmGateway = {
    getUserDraftNSABookingsIfExist: jest.fn(),
    updateNSABookings: jest.fn(),
    updateZeroCostBooking: jest.fn(),
    updateBookingStatus: jest.fn(),
  };

  const mockNsaBookingDetails: NsaBookingDetail[] = [
    {
      bookingId: '123',
      nsaStatus: CRMNsaStatus.AwaitingCscResponse,
      origin: CRMOrigin.CitizenPortal,
    },
    {
      bookingId: '321',
      nsaStatus: CRMNsaStatus.AwaitingCscResponse,
      origin: CRMOrigin.CitizenPortal,
    },
  ];

  const mockReceiptReference = 'my_receipt_reference';
  const mockBookingRef = '123456-123';
  const mockBookingProductRef = '123456-123-01';
  const mockReservationId = '456789-321';
  const mockBookingId = '1234-5678-9123-4567';
  const mockEmail = 'mock@email.com';
  const mockCandidateId = 'mock-candidate-id';
  const mockLicenceId = 'mockLicenceId';
  const mockPersonReference = '123456789';
  const mockTarget = Target.GB;
  const mockRemit = CRMRemit.England;
  const mockTestType = TestType.CAR;
  const mockDateTime = '2020-10-01';
  const mockBookingProductId = 'mock-booking-product-id';
  const mockRegion = TCNRegion.A;
  const mockPaymentId = 'AACC941E-D6CB-4AA0-92E8-FCADA53B7D01';

  beforeEach(() => {
    mockPaymentGateway.confirmCardPaymentIsComplete.mockResolvedValue({
      code: PaymentStatus.SUCCESS,
      receiptReference: mockReceiptReference,
      paymentId: mockPaymentId,
    });

    mockBookingService.completeBooking.mockResolvedValue({
      isConfirmed: true,
      lastRefundDate: '2021-04-01',
    });

    controller = new PaymentConfirmationController(
      mockNotificationsGateway as any,
      mockBookingService as any,
      mockPaymentGateway,
      mockCrmGateway as any,
    );

    req = {
      session: {
        candidate: {
          email: mockEmail,
          candidateId: mockCandidateId,
          personReference: mockPersonReference,
          licenceId: mockLicenceId,
        },
        currentBooking: {
          testType: mockTestType,
          dateTime: mockDateTime,
          receiptReference: mockReceiptReference,
          bookingRef: mockBookingRef,
          bookingProductRef: mockBookingProductRef,
          bookingId: mockBookingId,
          reservationId: mockReservationId,
          centre: {
            remit: mockRemit,
            region: mockRegion,
          },
          bookingProductId: mockBookingProductId,
          priceList: {
            testType: TestType.CAR,
            price: 11,
            product: {
              productId: '123',
              parentId: '321',
            },
          },
        },
        target: mockTarget,
      },
    };

    res = {
      redirect: jest.fn(),
      render: jest.fn(),
    };
  });

  afterEach(() => {
    jest.resetAllMocks();
  });

  describe('GET request', () => {
    test('redirects to the booking confirmation page', async () => {
      await controller.get(req, res);

      expect(logger.event).toHaveBeenCalledWith(BusinessTelemetryEvents.PAYMENT_BACK, expect.any(String));
      expect(res.redirect).toHaveBeenCalledWith('/booking-confirmation');
    });

    test('calls payment API to confirm the payment completion', async () => {
      await controller.get(req, res);

      expect(mockStore.reset).not.toHaveBeenCalled();
      expect(mockPaymentGateway.confirmCardPaymentIsComplete).toHaveBeenCalledWith(mockReceiptReference, mockCandidateId, mockPersonReference);
    });

    test('for compensation bookings, calls payment api to transfer finance transactions', async () => {
      req.session.currentBooking.compensationBooking = {
        ftts_bookingstatus: 675030008,
        ftts_selected: true,
        bookingId: '42bcf7e8-43fc-eb11-94ef-000d3ad6739c',
        bookingProductReference: 'B-000-193-849-01',
        bookingReference: 'B-000-193-849',
        productNumber: CRMProductNumber.CAR,
        compensationAssigned: '2021-08-13T15:00:02.000Z',
        candidateName: 'Wendy',
        bookingStatus: 675030008,
        candidateId: '4ddab30e-f90e-eb11-a813-000d3a7f128d',
        bookingProductId: 'mockOldBookingProductId',
      };

      await controller.get(req, res);

      expect(mockPaymentGateway.compensateBooking).toHaveBeenCalledWith(
        {
          compensatedBookingProductId: 'mockOldBookingProductId',
          newBookingProductId: 'mock-booking-product-id',
        },
        'mock-candidate-id',
        '123456789',
      );
    });

    test('throws an error when booking compensation is unsuccessful', async () => {
      req.session.currentBooking.compensationBooking = {
        ftts_bookingstatus: 675030008,
        ftts_selected: true,
        bookingId: '42bcf7e8-43fc-eb11-94ef-000d3ad6739c',
        bookingProductReference: 'B-000-193-849-01',
        bookingReference: 'B-000-193-849',
        productNumber: CRMProductNumber.CAR,
        compensationAssigned: '2021-08-13T15:00:02.000Z',
        candidateName: 'Wendy',
        bookingStatus: 675030008,
        candidateId: '4ddab30e-f90e-eb11-a813-000d3a7f128d',
        bookingProductId: 'mockOldBookingProductId',
      };
      const expectedError = new Error('compensation usuccessful');
      mockPaymentGateway.compensateBooking.mockRejectedValue(expectedError);

      await expect(controller.get(req, res)).rejects.toThrow(expectedError);
    });

    test('unreserves the booking and renders the payment unsuccessful page if the payment was not successful (call resolves)', async () => {
      const error: Partial<PaymentCandidate.CardPaymentCompletionResponse> = {
        code: PaymentStatus.FAILED,
        message: 'Payment unsuccessful',
      };
      const {
        bookingProductRef, bookingProductId, bookingId, reservationId,
      } = req.session.currentBooking;
      const { region } = req.session.currentBooking.centre;

      mockPaymentGateway.confirmCardPaymentIsComplete.mockResolvedValue(error as PaymentCandidate.CardPaymentCompletionResponse);

      await controller.get(req, res);

      expect(logger.warn).toHaveBeenCalledWith(
        'PaymentConfirmationController::confirmPayment: Payment status NOT successful - cancelling booking',
        {
          cpmsCode: 802,
          cpmsMessage: 'Payment unsuccessful',
          ...getCreatedBookingIdentifiers(req),
        },
      );
      expect(mockBookingService.unreserveBooking).toHaveBeenCalledWith(bookingProductRef, region, bookingProductId, bookingId, reservationId, CRMBookingStatus.Draft);
      expect(logger.event).toHaveBeenCalledWith(BusinessTelemetryEvents.PAYMENT_FAILED, expect.any(String), {
        cpmsCode: 802,
        ...getCreatedBookingIdentifiers(req),
      });
      expect(res.render).toHaveBeenCalledWith(PageNames.PAYMENT_UNSUCCESSFUL, { startAgainLink: '/choose-support?target=gb&lang=gb' });
    });

    test('unreserves the booking and renders the payment unsuccessful page if the payment was not successful (call fails)', async () => {
      const error = {
        response: {
          data: {
            code: PaymentStatus.FAILED,
          },
        },
      };
      const {
        bookingProductRef, bookingProductId, bookingId, reservationId,
      } = req.session.currentBooking;
      const { region } = req.session.currentBooking.centre;

      mockPaymentGateway.confirmCardPaymentIsComplete.mockRejectedValue(error);

      await controller.get(req, res);

      expect(mockBookingService.unreserveBooking).toHaveBeenCalledWith(bookingProductRef, region, bookingProductId, bookingId, reservationId, CRMBookingStatus.Draft);
      expect(logger.event).toHaveBeenCalledWith(BusinessTelemetryEvents.PAYMENT_FAILED, expect.any(String), {
        error,
        cpmsCode: PaymentStatus.FAILED,
        ...getCreatedBookingIdentifiers(req),
      });
      expect(res.render).toHaveBeenCalledWith(PageNames.PAYMENT_UNSUCCESSFUL, { startAgainLink: '/choose-support?target=gb&lang=gb' });
      expect(logger.error).toHaveBeenCalledWith(error, 'PaymentConfirmationController::confirmPayment: Error confirming payment for booking', {
        response: error.response?.data,
        cpmsCode: PaymentStatus.FAILED,
        ...getCreatedBookingIdentifiers(req),
      });
    });

    test('unreserves the booking and renders the payment error page if the payment call fails with an unknown error', async () => {
      const mockError = new Error('Payment api error');
      const {
        bookingProductRef, bookingProductId, bookingId, reservationId,
      } = req.session.currentBooking;
      const { region } = req.session.currentBooking.centre;

      mockPaymentGateway.confirmCardPaymentIsComplete.mockRejectedValue(mockError);

      await controller.get(req, res);

      expect(mockBookingService.unreserveBooking).toHaveBeenCalledWith(bookingProductRef, region, bookingProductId, bookingId, reservationId, CRMBookingStatus.Draft);
      expect(logger.event).toHaveBeenCalledWith(BusinessTelemetryEvents.PAYMENT_FAILED, expect.any(String), {
        error: mockError,
        ...getCreatedBookingIdentifiers(req),
      });
      expect(mockStore.reset).toHaveBeenCalledWith(req, res);
      expect(res.render).toHaveBeenCalledWith(PageNames.PAYMENT_CONFIRMATION_ERROR, { startAgainLink: '/choose-support?target=gb&lang=gb' });
      expect(logger.error).toHaveBeenCalledWith(mockError, 'PaymentConfirmationController::confirmPayment: Error confirming payment for booking', {
        response: undefined,
        ...getCreatedBookingIdentifiers(req),
      });
    });

    test('unreserves the booking and renders the booking cancelled page if the payment was cancelled by the user', async () => {
      req.session.currentBooking.paymentId = 'payment-id';
      const error = {
        response: {
          data: {
            code: PaymentStatus.USER_CANCELLED,
          },
        },
      };
      const {
        bookingProductRef, bookingProductId, bookingId, reservationId,
      } = req.session.currentBooking;
      const { region } = req.session.currentBooking.centre;

      mockPaymentGateway.confirmCardPaymentIsComplete.mockRejectedValue(error);

      await controller.get(req, res);

      expect(mockBookingService.unreserveBooking).toHaveBeenCalledWith(bookingProductRef, region, bookingProductId, bookingId, reservationId, CRMBookingStatus.AbandonedNonRecoverable, 'payment-id');
      expect(logger.event).toHaveBeenCalledWith(BusinessTelemetryEvents.PAYMENT_FAILED, expect.any(String), {
        error,
        cpmsCode: PaymentStatus.USER_CANCELLED,
        ...getCreatedBookingIdentifiers(req),
      });
      expect(logger.event).toHaveBeenCalledWith(BusinessTelemetryEvents.PAYMENT_CANCEL, expect.any(String), {
        error,
        cpmsCode: PaymentStatus.USER_CANCELLED,
        ...getCreatedBookingIdentifiers(req),
      });
      expect(mockStore.reset).toHaveBeenCalledWith(req, res);
      expect(res.render).toHaveBeenCalledWith(PageNames.PAYMENT_BOOKING_CANCELLED, { startAgainLink: '/choose-support?target=gb&lang=gb' });
      expect(logger.error).toHaveBeenCalledWith(error, 'PaymentConfirmationController::confirmPayment: Error confirming payment for booking', {
        response: error.response?.data,
        cpmsCode: PaymentStatus.USER_CANCELLED,
        ...getCreatedBookingIdentifiers(req),
      });
    });

    test('unreserves the booking and renders the booking cancelled page if the payment was cancelled by the user (no paymentId found)', async () => {
      req.session.currentBooking.paymentId = undefined;
      const error = {
        response: {
          data: {
            code: PaymentStatus.USER_CANCELLED,
          },
        },
      };
      const {
        bookingProductRef, bookingProductId, bookingId, reservationId,
      } = req.session.currentBooking;
      const { region } = req.session.currentBooking.centre;

      mockPaymentGateway.confirmCardPaymentIsComplete.mockRejectedValue(error);

      await controller.get(req, res);

      expect(mockBookingService.unreserveBooking).toHaveBeenCalledWith(bookingProductRef, region, bookingProductId, bookingId, reservationId, CRMBookingStatus.AbandonedNonRecoverable, undefined);
      expect(logger.error).toHaveBeenCalled();
      expect(logger.event).toHaveBeenCalledWith(BusinessTelemetryEvents.PAYMENT_FAILED, expect.any(String), {
        error,
        cpmsCode: PaymentStatus.USER_CANCELLED,
        ...getCreatedBookingIdentifiers(req),
      });
      expect(logger.event).toHaveBeenCalledWith(BusinessTelemetryEvents.PAYMENT_CANCEL, expect.any(String), {
        error,
        cpmsCode: PaymentStatus.USER_CANCELLED,
        ...getCreatedBookingIdentifiers(req),
      });
      expect(mockStore.reset).toHaveBeenCalledWith(req, res);
      expect(res.render).toHaveBeenCalledWith(PageNames.PAYMENT_BOOKING_CANCELLED, { startAgainLink: '/choose-support?target=gb&lang=gb' });
      expect(logger.error).toHaveBeenCalledWith(error, 'PaymentConfirmationController::confirmPayment: Error confirming payment for booking', {
        response: error.response?.data,
        cpmsCode: PaymentStatus.USER_CANCELLED,
        ...getCreatedBookingIdentifiers(req),
      });
    });

    test('updates booking status to \'system error non recoverable\' if the payment API throws a gateway error from cpms (error code 810)', async () => {
      req.session.currentBooking.paymentId = undefined;
      const error = {
        response: {
          data: {
            code: PaymentStatus.GATEWAY_ERROR,
          },
        },
      };
      const {
        bookingProductRef, bookingProductId, bookingId, reservationId,
      } = req.session.currentBooking;
      const { region } = req.session.currentBooking.centre;

      mockPaymentGateway.confirmCardPaymentIsComplete.mockRejectedValue(error);

      await controller.get(req, res);

      expect(mockStore.reset).toHaveBeenCalledWith(req, res);
      expect(mockBookingService.unreserveBooking).toHaveBeenCalledWith(bookingProductRef, region, bookingProductId, bookingId, reservationId, CRMBookingStatus.SystemErrorNonRecoverable);
      expect(res.render).toHaveBeenCalledWith(PageNames.PAYMENT_UNSUCCESSFUL, {
        startAgainLink: '/choose-support?target=gb&lang=gb',
      });
      expect(logger.event).toHaveBeenCalledWith(BusinessTelemetryEvents.PAYMENT_FAILED, expect.any(String), {
        error,
        cpmsCode: PaymentStatus.GATEWAY_ERROR,
        ...getCreatedBookingIdentifiers(req),
      });
      expect(logger.event).toHaveBeenCalledWith(BusinessTelemetryEvents.PAYMENT_GATEWAY_ERROR, expect.any(String), {
        error,
        cpmsCode: PaymentStatus.GATEWAY_ERROR,
        ...getCreatedBookingIdentifiers(req),
      });
    });

    test('updates booking status to \'system error non recoverable\' if the payment API throws a system error from cpms (error code 828)', async () => {
      req.session.currentBooking.paymentId = undefined;
      const error = {
        response: {
          data: {
            code: PaymentStatus.SYSTEM_ERROR,
          },
        },
      };
      const {
        bookingProductRef, bookingProductId, bookingId, reservationId,
      } = req.session.currentBooking;
      const { region } = req.session.currentBooking.centre;

      mockPaymentGateway.confirmCardPaymentIsComplete.mockRejectedValue(error);

      await controller.get(req, res);

      expect(mockBookingService.unreserveBooking).toHaveBeenCalledWith(bookingProductRef, region, bookingProductId, bookingId, reservationId, CRMBookingStatus.SystemErrorNonRecoverable);
      expect(mockStore.reset).toHaveBeenCalledWith(req, res);
      expect(res.render).toHaveBeenCalledWith(PageNames.PAYMENT_SYSTEM_ERROR, {
        startAgainLink: '/choose-support?target=gb&lang=gb',
        bookingRef: '123456-123',
      });
      expect(logger.event).toHaveBeenCalledWith(BusinessTelemetryEvents.PAYMENT_SYSTEM_ERROR, expect.any(String), {
        error,
        cpmsCode: PaymentStatus.SYSTEM_ERROR,
        ...getCreatedBookingIdentifiers(req),
      });
      expect(logger.event).toHaveBeenCalledWith(BusinessTelemetryEvents.PAYMENT_SYSTEM_ERROR, expect.any(String), {
        error,
        cpmsCode: PaymentStatus.SYSTEM_ERROR,
        ...getCreatedBookingIdentifiers(req),
      });
    });

    test('calls the booking service to confirm the booking', async () => {
      mockBookingService.completeBooking.mockResolvedValue({
        isConfirmed: true,
        lastRefundDate: '2021-04-01',
      });

      const currentBooking = { ...req.session.currentBooking };

      await controller.get(req, res);

      expect(mockBookingService.completeBooking).toHaveBeenCalledWith(
        req.session.candidate,
        currentBooking,
        mockPaymentId,
      );
      expect(res.redirect).toHaveBeenCalled();
    });

    test('resets the session and renders the slot confirmation error page if the slot confirmation fails', async () => {
      mockBookingService.completeBooking.mockResolvedValue({
        isConfirmed: false,
        bookingRef: mockBookingRef,
      });

      await controller.get(req, res);

      expect(mockBookingService.completeBooking).toHaveBeenCalledWith(req.session.candidate, req.session.currentBooking, mockPaymentId);
      expect(mockStore.reset).toHaveBeenCalledWith(req, res);
      expect(res.render).toHaveBeenCalledWith(PageNames.SLOT_CONFIRMATION_ERROR, {
        bookingReference: mockBookingRef,
        startAgainLink: expect.any(String),
      });
    });

    test('resets the session and renders the slot confirmation error page if completeBooking throws an error due to issue with tcn', async () => {
      const error = new Error('error completing booking');
      mockBookingService.completeBooking.mockRejectedValue(error);

      await controller.get(req, res);

      expect(logger.error).toHaveBeenCalledWith(error, 'payment-confirmation::get: Error completing booking', {
        bookingRef: mockBookingRef,
        bookingId: mockBookingId,
        bookingProductId: mockBookingProductId,
        bookingProductRef: mockBookingProductRef,
        candidateId: mockCandidateId,
        licenceId: mockLicenceId,
      });
      expect(mockStore.reset).toHaveBeenCalledWith(req, res);
      expect(res.render).toHaveBeenCalledWith(PageNames.SLOT_CONFIRMATION_ERROR, {
        bookingReference: mockBookingRef,
        startAgainLink: expect.any(String),
      });
    });

    test('if completing the booking fails due to a crm failure, we log and throw the error', async () => {
      const error = new CrmServerError('error with crm failing the completion of the booking');
      mockBookingService.completeBooking.mockRejectedValue(error);

      await expect(controller.get(req, res)).rejects.toThrow(error);
      expect(logger.error).toHaveBeenCalledWith(error, 'payment-confirmation::get: Error completing booking', {
        bookingRef: mockBookingRef,
        bookingId: mockBookingId,
        bookingProductId: mockBookingProductId,
        bookingProductRef: mockBookingProductRef,
        candidateId: mockCandidateId,
        licenceId: mockLicenceId,
      });
    });

    test('calls notification API to send a booking confirmation email', async () => {
      await controller.get(req, res);

      expect(mockNotificationsGateway.sendEmail).toHaveBeenCalledWith(mockEmail, mockEmailContent, mockBookingRef, mockTarget);
    });

    test('logs and swallows the error if the notifications call fails', async () => {
      const mockError = new Error('notifications error');
      mockNotificationsGateway.sendEmail.mockRejectedValue(mockError);

      await controller.get(req, res);

      expect(mockNotificationsGateway.sendEmail).toHaveBeenCalled();
      expect(logger.error).toHaveBeenCalledWith(Error('notifications error'), 'PaymentConfirmationController::get: Error sending booking confirmation email', {
        ...getCreatedBookingIdentifiers(req),
      });
    });

    test('changes booking status and nsa status of draft nsa bookings', async () => {
      mockCrmGateway.getUserDraftNSABookingsIfExist.mockResolvedValue(mockNsaBookingDetails);

      await controller.get(req, res);

      expect(mockCrmGateway.getUserDraftNSABookingsIfExist).toHaveBeenCalledWith(req.session.candidate.candidateId, req.session.currentBooking.testType as TestType);

      expect(mockCrmGateway.updateNSABookings).toHaveBeenCalledWith(mockNsaBookingDetails);
      expect(res.redirect).toHaveBeenCalledWith('/booking-confirmation');
    });

    test('error getUserDraftNSABookingsIfExist when trying change booking status and nsa status of draft nsa bookings', async () => {
      const mockError = new Error('update nsa bookings error');

      mockCrmGateway.getUserDraftNSABookingsIfExist.mockRejectedValue(mockError);

      await controller.get(req, res);

      expect(mockCrmGateway.getUserDraftNSABookingsIfExist).toHaveBeenCalled();
      expect(logger.error).toHaveBeenCalledWith(mockError, 'PaymentConfirmationController::get: Error executing batch request to update nsa booking status', { ...getCreatedBookingIdentifiers(req) });
    });

    test('error updateNSABookings when trying change booking status and nsa status of draft nsa bookings', async () => {
      mockCrmGateway.getUserDraftNSABookingsIfExist.mockResolvedValue(mockNsaBookingDetails);

      const mockError = new Error('update nsa bookings error');
      mockCrmGateway.updateNSABookings.mockRejectedValue(mockError);

      await controller.get(req, res);

      expect(mockCrmGateway.updateNSABookings).toHaveBeenCalled();
      expect(logger.error).toHaveBeenCalledWith(mockError, 'PaymentConfirmationController::get: Error executing batch request to update nsa booking status', { ...getCreatedBookingIdentifiers(req) });
    });

    test('for zero cost bookings do not calls payment API to confirm the payment completion', async () => {
      req.session.currentBooking.testType = TestType.ADIP1DVA;

      await controller.get(req, res);

      expect(mockPaymentGateway.confirmCardPaymentIsComplete).toHaveBeenCalledTimes(0);
      expect(mockCrmGateway.updateZeroCostBooking).toHaveBeenCalledTimes(1);
      expect(mockCrmGateway.updateZeroCostBooking).toHaveBeenCalledWith(mockBookingId);
    });

    describe('missing session data', () => {
      test('throws if missing currentBooking', async () => {
        req.session.currentBooking = undefined;

        await expect(controller.get(req, res)).rejects.toThrow();
      });

      test('throws if missing candidate', async () => {
        req.session.candidate = undefined;

        await expect(controller.get(req, res)).rejects.toThrow();
      });

      test('throws if missing payment data', async () => {
        req.session.candidate.receiptReference = undefined;
        req.session.candidate.candidateId = undefined;
        req.session.candidate.personReference = undefined;

        await expect(controller.get(req, res)).rejects.toThrow();
      });

      test('throws if missing email data', async () => {
        req.session.currentBooking.dateTime = undefined;
        req.session.currentBooking.centre = undefined;
        req.session.currentBooking.bookingRef = undefined;

        await expect(controller.get(req, res)).rejects.toThrow();
      });
    });
  });
});
